分类管理 删除分类

+++ mall/src/main/java/com/geekaca/mall/mapper/GoodsCategoryMapper.java
@@ -34,5 +34,5 @@ public interface GoodsCategoryMapper {
 
     GoodsCategory selectByCategoryName(String categoryName);
 
-    int deleteByIds(Integer[] ids);
+    int deleteByIds(Long[] ids);
 }
