+++ mall/src/main/java/com/geekaca/mall/service/impl/GoodsCategoryServiceImpl.java
@@ -3,7 +3,6 @@ package com.geekaca.mall.service.impl;
 import com.geekaca.mall.controller.admin.param.CategoryParam;
 import com.geekaca.mall.domain.GoodsCategory;
 import com.geekaca.mall.mapper.GoodsCategoryMapper;
-import com.geekaca.mall.mapper.GoodsInfoMapper;
 import com.geekaca.mall.service.GoodsCategoryService;
 import com.geekaca.mall.utils.PageQueryUtil;
 import com.geekaca.mall.utils.PageResult;
@@ -38,7 +37,7 @@ public class GoodsCategoryServiceImpl implements GoodsCategoryService {
     }
 
     @Override
-    public Boolean deleteCategory(Integer[] ids) {
+    public Boolean deleteCategory(Long[] ids) {
         if (ids.length < 1) {
             return false;
         }
